Added delete validation

+++ sps-server/src/main/Java/com/azry/sps/server/services/user/SystemUserManagerBean.java
@@ -126,7 +126,9 @@ public class SystemUserManagerBean implements SystemUserManager {
 	@Override
 	public void remove(long id) {
 		SystemUser user = em.find(SystemUser.class, id);
-		em.remove(user);
+		if (user != null) {
+			em.remove(user);
+		}
 	}
 
 	@Override
